Initial Upload to GIT of sample code
 ServerCommunicationManager package, was used inside a photo book
creation and editor application. All communication from within the
application to and from the server needed to be managed. This package
allowed the application to always verify that the users session was
valid, the user was who the server expected them to be, and to
synchronously execute server requests depending on the particulars of
the user and the application.  It also allowed events and functions and
actions from within the application to be prepped, stored and processed
or not depending on the server communication. This way the user would
never edit, or save locally without the actual changes being accepted
on the server.